chore: Create all test data using rollup
