+++ scripts/rollup.config.js
@@ -1,6 +1,7 @@
 const generate = require('videojs-generate-rollup-config');
 const worker = require('@gkatsev/rollup-plugin-bundle-worker');
 const {terser} = require('rollup-plugin-terser');
+const createTestData = require('./create-test-data.js');
 
 // see https://github.com/videojs/videojs-generate-rollup-config
 // for options
@@ -27,6 +28,7 @@ const options = {
     defaults.module.splice(2, 0, 'worker');
     defaults.browser.splice(2, 0, 'worker');
     defaults.test.splice(3, 0, 'worker');
+    defaults.test.splice(0, 0, 'createTestData');
 
     // istanbul is only in the list for regular builds and not watch
     if (defaults.test.indexOf('istanbul') !== -1) {
@@ -42,7 +44,8 @@ const options = {
         output: {comments: 'some'},
         compress: {passes: 2},
         include: [/^.+\.min\.js$/]
-      })
+      }),
+      createTestData: createTestData()
     });
   },
   babel(defaults) {
