+++ test/master-playlist-controller.test.js
@@ -11,7 +11,7 @@ import {
   setupMediaSource,
   downloadProgress
 } from './test-helpers.js';
-import manifests from './dist/test-manifests.js';
+import manifests from 'create-test-data!manifests';
 import {
   MasterPlaylistController,
   DEFAULT_AUDIO_CODEC,
@@ -39,7 +39,7 @@ import {
   mp4Video as mp4VideoSegment,
   mp4AudioInit as mp4AudioInitSegment,
   mp4Audio as mp4AudioSegment
-} from './dist/test-segments';
+} from 'create-test-data!segments';
 import {
   timeRangesEqual,
   bandwidthWithinTolerance
